svd-parser 0.10 update

+++ src/patch/yaml/yaml_parser.rs
@@ -102,7 +102,7 @@ pub struct PeripheralBody {
     pub display_name: Option<String>,
     pub description: Option<String>,
     pub group_name: Option<String>,
-    pub base_address: Option<u32>,
+    pub base_address: Option<u64>,
 }
 
 #[derive(Debug, Deserialize, Clone)]
@@ -212,8 +212,8 @@ pub struct Device {
 #[serde(deny_unknown_fields)]
 pub struct RegisterProperties {
     pub size: Option<u32>,
-    pub reset_value: Option<u32>,
-    pub reset_mask: Option<u32>,
+    pub reset_value: Option<u64>,
+    pub reset_mask: Option<u64>,
     pub access: Option<Access>,
 }
